refactor: extract environments code into features
