+++ packages/frontend/editor-ui/src/Interface.ts
@@ -1013,12 +1013,6 @@ export type NodeAuthenticationOption = {
 	displayOptions?: IDisplayOptions;
 };
 
-export interface EnvironmentVariable {
-	id: string;
-	key: string;
-	value: string;
-}
-
 export type ExecutionFilterMetadata = {
 	key: string;
 	value: string;
