+++ packages/frontend/editor-ui/src/__tests__/server/endpoints/variable.ts
@@ -2,7 +2,7 @@ import type { Request, Server } from 'miragejs';
 import { Response } from 'miragejs';
 import type { AppSchema } from '../types';
 import { jsonParse } from 'n8n-workflow';
-import type { EnvironmentVariable } from '@/Interface';
+import type { EnvironmentVariable } from '@/features/environments.ee/environments.types';
 
 export function routesForVariables(server: Server) {
 	server.get('/rest/variables', (schema: AppSchema) => {
