+++ packages/frontend/editor-ui/src/__tests__/server/models/variable.ts
@@ -1,4 +1,4 @@
-import type { EnvironmentVariable } from '@/Interface';
+import type { EnvironmentVariable } from '@/features/environments.ee/environments.types';
 import { Model } from 'miragejs';
 import type { ModelDefinition } from 'miragejs/-types';
