+++ packages/frontend/editor-ui/src/components/CodeNodeEditor/completions/workflow.completions.ts
@@ -1,4 +1,4 @@
-import { addVarType } from '../utils';
+import { addVarType } from '@/features/environments.ee/completions/variables.completions';
 import type { Completion, CompletionContext, CompletionResult } from '@codemirror/autocomplete';
 import { useI18n } from '@n8n/i18n';
