+++ packages/frontend/editor-ui/src/features/environments.ee/components/VariablesUsageBadge.vue
