+++ packages/frontend/editor-ui/src/features/environments.ee/environments.types.ts
@@ -1,0 +1,5 @@
+export interface EnvironmentVariable {
+	id: string;
+	key: string;
+	value: string;
+}
