+++ packages/frontend/editor-ui/src/features/environments.ee/views/VariablesView.test.ts
@@ -1,8 +1,8 @@
-import VariablesView from '@/views/VariablesView.vue';
+import VariablesView from './VariablesView.vue';
 import { useSettingsStore } from '@/stores/settings.store';
 import { useUsersStore } from '@/stores/users.store';
 import { useRBACStore } from '@/stores/rbac.store';
-import { useEnvironmentsStore } from '@/stores/environments.ee.store';
+import { useEnvironmentsStore } from '../environments.store';
 import { createComponentRenderer } from '@/__tests__/render';
 import { EnterpriseEditionFeature } from '@/constants';
 import { STORES } from '@n8n/stores';
@@ -11,7 +11,7 @@ import { mockedStore, SETTINGS_STORE_DEFAULT_STATE } from '@/__tests__/utils';
 import { createRouter, createWebHistory } from 'vue-router';
 import userEvent from '@testing-library/user-event';
 import { waitFor, within } from '@testing-library/vue';
-import type { EnvironmentVariable } from '@/Interface';
+import type { EnvironmentVariable } from '../environments.types';
 import type { IUser } from '@n8n/rest-api-client/api/users';
 import type { Scope } from '@n8n/permissions';
